WP Explorer: src/my-wordpress/ deleted — the app owns its parts, the shell keeps four shared modules
The port to the App Framework left nine files behind under
src/my-wordpress/. Each goes where its importers are:

  - agents-face / agents-rest / agents-send-to → apps/my-wordpress/parts/
    (only the app ever imported them)
  - types.ts (609 lines) → deleted; the one symbol still used,
    `UserFootprint`, now lives in the app's parts/types.ts
  - agents-types.ts → src/agents-types.ts, beside agents-dispatch and
    agents-chat-store, the shell modules that share it with the app
  - explorer-open / footprint-target → src/open-targets/, the
    cross-bundle "open the app on X" hand-offs the desktop tiles,
    the preview pane, the Corkboard and the iframe bridge call
  - rest-trash → src/desktop-files/, beside its only consumer
  - integrations/woocommerce.ts → src/plugins/my-wordpress-woocommerce/,
    a plugin-shaped companion bundle like the widgets and wallpapers

The i18n handle map gains the my-wordpress, code-blue and WooCommerce
prefixes so their strings reach the handles that set translations.
Docs, comments and the DEVELOPMENT.md tree follow.
