+++ src/desktop-files/rest-trash.ts
