+++ src/plugins/my-wordpress-woocommerce/index.ts
